Реализует отрисовку цитируемых фильмов и фильмов с самым высоким рейтингом
